Move Bai2::BaiFile to expected location
